의사 추가/전체 조회 기능 수정, 사용자 전체 조회 기능 추가

+++ src/main/java/com/project/doctor_fish_back/dto/response/review/RespGetReviewListDto.java
@@ -10,5 +10,5 @@ import java.util.List;
 @Data
 public class RespGetReviewListDto {
     private List<Review> reviews;
-    private int reviewCount;
+    private Long reviewCount;
 }
